+++ QUANTAXIS/__init__.py
@@ -33,7 +33,7 @@ by yutiansut
 """
 
 
-__version__ = ' 0.4.0-alpha-dev05'
+__version__ = ' 0.4.0-alpha-dev07'
 __author__ = 'yutiansut'
 
 # fetch methods
@@ -81,10 +81,11 @@ from QUANTAXIS.QAUtil import (QA_util_sql_mongo_setting, QA_util_cfg_initial, QA
                               QA_util_log_debug, QA_util_log_expection, QA_util_log_info,
                               QA_start_initial, QA_Setting, QA_util_get_date_index,
                               QA_util_get_index_date, QA_util_get_real_date, QA_util_select_hours,
-                              QA_util_select_min, QA_util_time_delay)
+                              QA_util_select_min, QA_util_time_delay,QA_util_time_now)
 
 from QUANTAXIS.QAIndicator import *
 from QUANTAXIS.QASQL import qasql, qacold
+from QUANTAXIS.QAWeb import QA_Web
 # CMD and Cli
 import QUANTAXIS.QACmd
