--- QUANTAXISWebkit/pyBackend/readme.md
@@ -1,3 +1,0 @@
-# pyBackend
-
-准备引入一个基于flask的server,负责替代部分原先nodejs的服务器功能
