+++ requirements.txt
@@ -3,5 +3,7 @@ celery
 pandas
 lxml
 requests
-numpy 
+numpy==1.12.0
 pytesseract
+flask-sockets==0.2.1 
+gevent-websocket==0.10.1
