+++ setup.py
@@ -21,7 +21,7 @@ NAME = "quantaxis"
 """
 名字，一般放你包的名字即可
 """
-PACKAGES = ["QUANTAXIS", "QUANTAXIS.QAFetch", "QUANTAXIS.QACmd", "QUANTAXIS.QAMarket",
+PACKAGES = ["QUANTAXIS", "QUANTAXIS.QAFetch", "QUANTAXIS.QACmd", "QUANTAXIS.QAMarket",'QUANTAXIS.QAWeb',
             "QUANTAXIS.QABacktest", "QUANTAXIS.QASQL", "QUANTAXIS.QATask",  "QUANTAXIS.QASpider",
             "QUANTAXIS.QASU", "QUANTAXIS.QAUtil", "QUANTAXIS.QAARP", "QUANTAXIS.QASignal", "QUANTAXIS.QAIndicator"]
 """
@@ -63,12 +63,13 @@ setup(
         'Intended Audience :: Developers',
         'Operating System :: OS Independent',
     ],
-    install_requires=['pandas>=0.20', 'numpy==1.12.0', 'tushare==0.7.4',
+    install_requires=['pandas>=0.20', 'numpy==1.12.0', 'tushare==0.7.4','flask-sockets==0.2.1 ',
                       'pymongo>=3.4', 'celery>=4.0.0', 'six>=1.10.0', 'tabulate>=0.7.7',
                       'docopt>=0.6.2', 'zenlog>=1.1', 'delegator.py>=0.0.12','flask==0.12.2'],
     entry_points={
         'console_scripts': [
-            'quantaxis=QUANTAXIS.QACmd:QA_cmd'
+            'quantaxis=QUANTAXIS.QACmd:QA_cmd',
+            'quantaxis_web=QUANTAXIS.QAWeb.QA_Web:main'
         ]
     },
     keywords=KEYWORDS,
